separated seed data into separate file
